requirements/common.txt
# python-telegram-bot
# httpx
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.5.30
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.9
pycparser==2.20
python-telegram-bot==12.7
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
tornado==6.0.4